src/requirements.txt
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.3
Markdown==3.3.7
psycopg2-binary==2.9.3
requests==2.26.0
pytest==6.2.5
pytest-django==4.4.0
celery==4.4.1
redis==3.4.1
django-environ==0.8.1
gunicorn
drf-yasg
django-phonenumber-field
phonenumbers
django-filter
djangorestframework-simplejwt
django-countries